moved consts to pkg
