+++ pkg/da/ipld/wrapper/nmt_wrapper.go
@@ -6,7 +6,7 @@ import (
 	"github.com/celestiaorg/nmt"
 	"github.com/celestiaorg/rsmt2d"
 
-	"github.com/celestiaorg/celestia-core/types/consts"
+	"github.com/celestiaorg/celestia-core/pkg/consts"
 )
 
 // Fulfills the rsmt2d.Tree interface and rsmt2d.TreeConstructorFn function
