+++ types/block.go
@@ -21,10 +21,10 @@ import (
 	tmmath "github.com/celestiaorg/celestia-core/libs/math"
 	"github.com/celestiaorg/celestia-core/libs/protoio"
 	tmsync "github.com/celestiaorg/celestia-core/libs/sync"
+	"github.com/celestiaorg/celestia-core/pkg/consts"
 	"github.com/celestiaorg/celestia-core/pkg/da/ipld/wrapper"
 	tmproto "github.com/celestiaorg/celestia-core/proto/tendermint/types"
 	tmversion "github.com/celestiaorg/celestia-core/proto/tendermint/version"
-	"github.com/celestiaorg/celestia-core/types/consts"
 	"github.com/celestiaorg/celestia-core/version"
 )
