+++ types/shares_test.go
@@ -13,7 +13,7 @@ import (
 
 	tmbytes "github.com/celestiaorg/celestia-core/libs/bytes"
 	"github.com/celestiaorg/celestia-core/libs/protoio"
-	"github.com/celestiaorg/celestia-core/types/consts"
+	"github.com/celestiaorg/celestia-core/pkg/consts"
 	"github.com/celestiaorg/nmt/namespace"
 	"github.com/celestiaorg/rsmt2d"
 	"github.com/stretchr/testify/assert"
